Rename remaining .jsx files to .js
