+++ src/app/components/collapsing-pane/collapsing-pane.js
