+++ src/app/components/content-group/content-group.js
