+++ src/app/components/quote/quote.js
