+++ src/app/components/shell/header/sticky-note/sticky-note.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import RawHTML from '~/components/jsx-helpers/raw-html';
-import {usePutAway} from '../../shared.jsx';
+import {usePutAway} from '../../shared.js';
 import './sticky-note.scss';
 
 export default function StickyNote({stickyData}) {
