+++ src/app/components/shell/shared.js
