+++ src/app/pages/errata-detail/progress-bar/progress-bar.js
