+++ src/app/pages/partners/partner-details/info-request-form/info-request-form.js
@@ -7,7 +7,7 @@ import MultiPageForm from '~/components/multi-page-form/multi-page-form';
 import BookTagsMultiselect, {BookTagsContextProvider} from '~/components/multiselect/book-tags/book-tags';
 import useSearchContext from '../../search-context';
 import FormRadiogroup from '~/components/form-radiogroup/form-radiogroup';
-import FormSelect from '~/components/form-select/form-select.jsx';
+import FormSelect from '~/components/form-select/form-select';
 import {FilteringSelect} from '~/components/form-elements/form-elements';
 import inputProps from '~/components/form-elements/input-props';
 import FormInput from '~/components/form-input/form-input';
