+++ src/app/pages/partners/partner-details/partner-details.js
@@ -5,7 +5,7 @@ import Carousel from './carousel/carousel';
 import Reviews from './reviews/reviews';
 import RawHTML from '~/components/jsx-helpers/raw-html';
 import TabGroup from '~/components/tab-group/tab-group';
-import ContentGroup from '~/components/content-group/content-group.jsx';
+import ContentGroup from '~/components/content-group/content-group';
 import booksPromise from '~/models/books';
 import analyticsEvents from '../analytics-events';
 import InfoRequestForm from './info-request-form/info-request-form';
