+++ src/app/pages/team/team.js
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 import LoaderPage from '~/components/jsx-helpers/loader-page';
 import Hero from '~/components/hero/hero';
 import TabGroup from '~/components/tab-group/tab-group';
-import ContentGroup from '~/components/content-group/content-group.jsx';
+import ContentGroup from '~/components/content-group/content-group';
 import AccordionGroup from '~/components/accordion-group/accordion-group.js';
 import PeopleTab from './people-tab/people-tab';
 import './team.scss';
